Support both bootstrap and update operations
bootstrap() method now takes in a parameter that specifies the 'call' web service name, just like the other methods. This allows it to access both the bootstrap and update web services.

Rename the sample IN file from 02-data.zip to 02-bootstrap.zip, and OUT file from 02-data.txt to 02-bootstrap.zip, to match this change.